requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
certifi==2021.5.30
charset-normalizer==2.0.4; python_version >= '3'
click==8.0.1; python_version >= '3.6'
flask==2.0.1
gunicorn==20.1.0
idna==3.2; python_version >= '3'
itsdangerous==2.0.1; python_version >= '3.6'
jinja2==3.0.1; python_version >= '3.6'
markupsafe==2.0.1; python_version >= '3.6'
marshmallow==3.13.0
python-dotenv==0.19.0
redis==3.5.3
requests==2.26.0
urllib3==1.26.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
werkzeug==2.0.1; python_version >= '3.6'
